LayoutTests/svg/dynamic-updates/SVGFEDistantLightElement-svgdom-elevation-prop-expected.txt

CONSOLE MESSAGE: line 50: TypeError: 'undefined' is not an object (evaluating 'distantLight.elevation.baseVal = 20')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'elevation' property of the SVGFEDistantLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL distantLight.elevation.baseVal should be 50. Threw exception TypeError: 'undefined' is not an object (evaluating 'distantLight.elevation.baseVal')
